complete notification feature

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -58,20 +58,47 @@ export async function POST(
     if (!loggedInUser) {
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
-    await prisma.like.upsert({
+
+    const post = await prisma.post.findUnique({
       where: {
-        userId_postId: {
+        id: postId,
+      },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId,
+          },
+        },
+        create: {
           userId: loggedInUser.id,
           postId,
         },
-      },
-      create: {
-        userId: loggedInUser.id,
-        postId,
-      },
-      update: {},
-    });
-
+        update: {},
+      }),
+      ...(post.userId !== loggedInUser.id
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                type: "LIKE",
+                postId,
+              },
+            }),
+          ]
+        : []),
+    ]);
     return new Response();
   } catch (error) {
     return Response.json({ error: "Internal server error" }, { status: 500 });
@@ -87,12 +114,36 @@ export async function DELETE(
     if (!loggedInUser) {
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
-    await prisma.like.deleteMany({
+
+    const post = await prisma.post.findUnique({
       where: {
-        userId: loggedInUser.id,
-        postId,
+        id: postId,
+      },
+      select: {
+        userId: true,
       },
     });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId,
+        },
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          type: "LIKE",
+          postId,
+        },
+      }),
+    ]);
     return new Response();
   } catch (error) {
     return Response.json({ error: "Internal server error" }, { status: 500 });
